requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-extensions==3.2.3
flake8==6.0.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
